Extract repro for #12720 into a separate file
