+++ frontend/test/metabase/scenarios/dashboard/dashboard.cy.spec.js
@@ -269,108 +269,6 @@ describe("scenarios > dashboard", () => {
     checkOptionsForFilter("Category");
   });
 
-  it.skip("should show QB question on a dashboard with filter connected to card without data-permission (metabase#12720)", () => {
-    // In this test we're using already present question ("Orders") and the dashboard with that question ("Orders in a dashboard")
-    cy.log(
-      "Add filter to the dashboard with the default value (after January 1st, 2020)",
-    );
-    cy.request("PUT", "/api/dashboard/1", {
-      parameters: [
-        {
-          default: "2020-01-01~",
-          id: "d3b78b27",
-          name: "Date Filter",
-          slug: "date_filter",
-          type: "date/all-options",
-        },
-      ],
-    });
-
-    cy.createNativeQuestion({
-      name: "12720_SQL",
-      native: {
-        query: "SELECT * FROM ORDERS WHERE {{filter}}",
-        "template-tags": {
-          filter: {
-            id: "1d006bb7-045f-6c57-e41b-2661a7648276",
-            name: "filter",
-            "display-name": "Filter",
-            type: "dimension",
-            dimension: ["field", ORDERS.CREATED_AT, null],
-            "widget-type": "date/month-year",
-            default: null,
-          },
-        },
-      },
-    }).then(({ body: { id: SQL_ID } }) => {
-      cy.log("Add SQL question to the dashboard");
-
-      cy.request("POST", "/api/dashboard/1/cards", {
-        cardId: SQL_ID,
-      }).then(({ body: { id: SQL_DASH_CARD_ID } }) => {
-        cy.log(
-          "Edit both cards (adjust their size and connect them to the filter)",
-        );
-
-        cy.request("PUT", "/api/dashboard/1/cards", {
-          cards: [
-            {
-              id: 1,
-              card_id: 1,
-              row: 0,
-              col: 0,
-              sizeX: 5,
-              sizeY: 5,
-              parameter_mappings: [
-                {
-                  parameter_id: "d3b78b27",
-                  card_id: 1,
-                  target: ["dimension", ["field", ORDERS.CREATED_AT, null]],
-                },
-              ],
-              visualization_settings: {},
-            },
-            {
-              id: SQL_DASH_CARD_ID,
-              card_id: SQL_ID,
-              row: 0,
-              col: 6, // previous card's sizeX + 1 (making sure they don't overlap)
-              sizeX: 5,
-              sizeY: 5,
-              parameter_mappings: [
-                {
-                  parameter_id: "d3b78b27",
-                  card_id: SQL_ID,
-                  target: ["dimension", ["template-tag", "filter"]],
-                },
-              ],
-              visualization_settings: {},
-            },
-          ],
-        });
-      });
-    });
-    cy.server();
-    cy.route("POST", "/api/card/*/query").as("cardQuery");
-
-    cy.signIn("nodata");
-
-    clickThrough("12720_SQL");
-    clickThrough("Orders");
-
-    /**
-     * Helper function related to this test only
-     */
-    function clickThrough(title) {
-      cy.visit("/dashboard/1");
-      cy.wait("@cardQuery.all");
-      cy.get(".LegendItem")
-        .contains(title)
-        .click();
-      cy.findByText(/^January 17, 2020/);
-    }
-  });
-
   it("should not get the parameter values from the field API", () => {
     // In this test we're using already present dashboard ("Orders in a dashboard")
     const FILTER_ID = "d7988e02";
